Refactor tests
- Use DiskStore in TCK tests.

- Put StoreContext & ResolvedStoreConfig in their own files.

- Share some stuff between JUnit 5 tests & the TCK.

- Some other pedantic refactorings.
