+++ methanol/src/test/java/com/github/mizosoft/methanol/testing/MockWebServerExtension.java
@@ -1,7 +1,31 @@
-package com.github.mizosoft.methanol;
+/*
+ * Copyright (c) 2019-2021 Moataz Abdelnasser
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package com.github.mizosoft.methanol.testing;
 
 import static com.github.mizosoft.methanol.testutils.TestUtils.localhostSslContext;
 
+import com.github.mizosoft.methanol.Methanol;
+import com.github.mizosoft.methanol.Methanol.Builder;
 import java.io.IOException;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -16,39 +40,24 @@ import okhttp3.mockwebserver.MockWebServer;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.junit.jupiter.api.extension.AfterAllCallback;
 import org.junit.jupiter.api.extension.AfterEachCallback;
-import org.junit.jupiter.api.extension.BeforeAllCallback;
-import org.junit.jupiter.api.extension.BeforeEachCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.api.extension.ExtensionContext.Namespace;
 import org.junit.jupiter.api.extension.ExtensionContext.Store.CloseableResource;
 import org.junit.jupiter.api.extension.ParameterContext;
 import org.junit.jupiter.api.extension.ParameterResolutionException;
 import org.junit.jupiter.api.extension.ParameterResolver;
+import org.junit.platform.commons.support.AnnotationSupport;
 
 /**
  * An extension that provides plain or secure {@code MockWebServers} and {@code Methanol.Builder}
  * either explicitly or by resolving parameters.
  */
-public final class MockWebServerProvider
-    implements BeforeAllCallback,
-        BeforeEachCallback,
-        AfterAllCallback,
-        AfterEachCallback,
-        ParameterResolver {
+public final class MockWebServerExtension
+    implements AfterAllCallback, AfterEachCallback, ParameterResolver {
   private static final Namespace EXTENSION_NAMESPACE =
-      Namespace.create(MockWebServerProvider.class);
+      Namespace.create(MockWebServerExtension.class);
 
-  private ManagedServers explicitServers;
-
-  @Override
-  public void beforeEach(ExtensionContext context) {
-    explicitServers = ManagedServers.get(context);
-  }
-
-  @Override
-  public void beforeAll(ExtensionContext context) {
-    explicitServers = ManagedServers.get(context);
-  }
+  public MockWebServerExtension() {}
 
   @Override
   public void afterEach(ExtensionContext context) throws Exception {
@@ -65,42 +74,28 @@ public final class MockWebServerProvider
       ParameterContext parameterContext, ExtensionContext extensionContext)
       throws ParameterResolutionException {
     var type = parameterContext.getParameter().getType();
-    return type == MockWebServer.class || type == Methanol.Builder.class;
+    return type == MockWebServer.class || type == Builder.class;
   }
 
   @Override
   public Object resolveParameter(
       ParameterContext parameterContext, ExtensionContext extensionContext)
       throws ParameterResolutionException {
-    // TODO BUG with UseHttps!
     var type = parameterContext.getParameter().getType();
     var servers = ManagedServers.get(extensionContext);
     var executable = parameterContext.getDeclaringExecutable();
-    boolean useHttps =
-        executable.isAnnotationPresent(UseHttps.class)
-            || extensionContext
-                .getElement()
-                .map(e -> e.isAnnotationPresent(UseHttps.class))
-                .orElse(false);
+    boolean useHttps = AnnotationSupport.isAnnotated(executable, UseHttps.class);
     if (type == MockWebServer.class) {
       try {
         return servers.newServer(executable, useHttps);
       } catch (IOException e) {
         throw new ParameterResolutionException("couldn't start server", e);
       }
-    } else if (type == Methanol.Builder.class) {
+    } else if (type == Builder.class) {
       return servers.newClientBuilder(executable, useHttps);
     } else {
       throw new UnsupportedOperationException("unsupported type: " + type.toString());
     }
-  }
-
-  public MockWebServer newServer(boolean useHttps) throws IOException {
-    var servers = explicitServers;
-    if (servers == null) {
-      throw new IllegalStateException("expected beforeAll or beforeEach to be called");
-    }
-    return servers.newServer(null, useHttps); // use null key for default context
   }
 
   @Retention(RetentionPolicy.RUNTIME)
@@ -120,7 +115,7 @@ public final class MockWebServerProvider
       return getContext(key).newServer(useHttps);
     }
 
-    Methanol.Builder newClientBuilder(@Nullable Object key, boolean useHttps) {
+    Builder newClientBuilder(@Nullable Object key, boolean useHttps) {
       return getContext(key).newClientBuilder(useHttps);
     }
 
@@ -129,8 +124,8 @@ public final class MockWebServerProvider
     }
 
     void shutdownAll() throws IOException {
-      for (var ctx : contextMap.values()) {
-        ctx.shutdownServers();
+      for (var context : contextMap.values()) {
+        context.shutdownServers();
       }
 
       contextMap.clear();
@@ -161,7 +156,7 @@ public final class MockWebServerProvider
         return server;
       }
 
-      Methanol.Builder newClientBuilder(boolean useHttps) {
+      Builder newClientBuilder(boolean useHttps) {
         return Methanol.newBuilder()
             .apply(
                 builder -> {
